# backend/api/swagger.yml
openapi: 3.0.3
info:
  title: University Platform API
  description: API для платформы отзывов об университетах. Включает аутентификацию, управление профилем, университетами и отзывами.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный сервер разработки
  - url: https://your-render-app.onrender.com/api/v1
    description: Продакшн сервер (Render)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]

    UpdateUserRequest:
      type: object
      required: [firstname, lastname]
      properties:
        firstname:
          type: string
        lastname:
          type: string

    RegisterRequest:
      type: object
      required: [firstname, lastname, email, password]
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    AuthenticationRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string

    UniversityResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        country:
          type: string
        city:
          type: string
        description:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        averageRating:
          type: number
          format: double
        tags:
          type: array
          items:
            type: string

    CreateUniversityRequest:
      type: object
      required: [name, city]
      properties:
        name:
          type: string
        country:
          type: string
        city:
          type: string
        description:
          type: string
        website:
          type: string
        logoUrl:
          type: string

    UpdateUniversityRequest:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        city:
          type: string
        description:
          type: string
        website:
          type: string
        logoUrl:
          type: string

    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        rating:
          type: integer
        tags:
          type: array
          items:
            type: string
        userName:
          type: string
        userId:
          type: integer
          format: int64
        universityId:
          type: integer
          format: int64
        universityName:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateReviewRequest:
      type: object
      required: [universityId, text, rating]
      properties:
        universityId:
          type: integer
          format: int64
        text:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        tags:
          type: array
          items:
            type: string

paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'

  /auth/authenticate:
    post:
      tags: [Authentication]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'

  /auth/me:
    get:
      tags: [Authentication]
      summary: Получить данные текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /users/profile:
    patch:
      tags: [Users]
      summary: Обновить профиль (только имя/фамилия)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /universities:
    get:
      tags: [Universities]
      summary: Список всех университетов
      responses:
        '200':
          description: Список университетов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UniversityResponse'
    post:
      tags: [Universities]
      summary: Создать университет (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUniversityRequest'
      responses:
        '200':
          description: Университет создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniversityResponse'

  /universities/{id}:
    get:
      tags: [Universities]
      summary: Получить университет по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
    put:
      tags: [Universities]
      summary: Обновить данные университета (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUniversityRequest'
      responses:
        '200':
          description: Данные обновлены
    delete:
      tags: [Universities]
      summary: Удалить университет (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Университет удален

  /reviews/university/{universityId}:
    get:
      tags: [Reviews]
      summary: Отзывы конкретного университета
      parameters:
        - name: universityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список отзывов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'

  /reviews:
    post:
      tags: [Reviews]
      summary: Добавить отзыв
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Отзыв добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'

  /reviews/{id}:
    delete:
      tags: [Reviews]
      summary: Удалить отзыв (Автор или Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Отзыв удален