# backend/src/main/java/dev/mathalama/backend/web/dto/UserResponse.java
package dev.mathalama.backend.web.dto;

import dev.mathalama.backend.domain.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class UserResponse {
    private Long id;
    private String email;
    private String firstname;
    private String lastname;
    private Role role;
}